requirements.txt
python-telegram-bot==20.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
